--- packages/common/sqlite-vfs-fdb/src/impls/mod.rs
@@ -1,1 +1,0 @@
-pub mod pages;
