--- packages/common/sqlite-vfs-fdb/src/impls/pages/mod.rs
@@ -1,3 +1,0 @@
-pub mod fdb;
-pub mod utils;
-pub mod vfs;
