+++ packages/common/sqlite-vfs-fdb/src/lib.rs
@@ -1,9 +1,12 @@
-pub mod impls;
+pub mod fdb;
 pub mod metrics;
 mod sqlite;
+pub mod utils;
+pub mod vfs;
 
 // Re-export foundationdb for convenience
 pub use foundationdb;
 
 // Re-export main functions from sqlite module
 pub use sqlite::{close_sqlite_db, execute_sql, open_sqlite_db, query_count};
+
