+++ packages/common/sqlite-vfs-fdb/src/utils.rs
